+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- `untracked()` - A function that reads the value of a signal without tracking it.
 - `structuralEqual()` - a structural equality function which uses `Object.is()` semantics.
 - `looseStructuralEqual()` - a structural equality function which uses `==` semantics.